requirements.txt
# This file contains all project dependencies
# It's a good practice to keep this file up-to-date
# You can generate this file using: pip freeze > requirements.txt

# Data Ingestion
kafka-python
boto3

# Data Processing
pyspark==3.1.2 # Specify compatible version for your EMR cluster
pyflink==1.15.0

# Model Training
tensorflow==2.9.0
horovod==0.28.0

# Model Serving
fastapi==0.95.0
uvicorn==0.21.1
redis==4.5.4
pydantic==1.10.7

# API
requests==2.28.2

# Orchestration
apache-airflow[amazon]==2.5.0

# Testing
pytest==7.3.1
pytest-cov==4.0.0

# Other
pandas==1.5.3
scikit-learn==1.2.2